streamer/requirements.txt
livekit-agents==0.2.0
python-dotenv==1.0.0
